--- src/app/admin/add-product/add-product.component.css
@@ -1,14 +1,0 @@
-label{
-    font-weight: 600;
-}
-
-.container{
-    margin-left: 50px;
-}
-
-.form .floated{
-    width: 45%;
-    margin-right:30px ;
-    float: left;
-}
-
