+++ src/app/admin/admin-routing.module.ts
@@ -1,6 +1,5 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { AddProductComponent } from './add-product/add-product.component';
 import { AdminHomeComponent } from './admin-home/admin-home.component';
 import { InventoryComponent } from './inventory/inventory.component';
 
@@ -11,10 +10,6 @@ const routes: Routes = [{
     {
       path:"",
       component:InventoryComponent
-    },
-    {
-      path:"addProduct",
-      component:AddProductComponent
     }
   ]
 }];
