+++ src/app/admin/inventory/inventory.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { ProductsService } from 'src/app/services/products.service';
 import * as $ from 'jquery'
+import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 @Component({
   selector: 'app-inventory',
   templateUrl: './inventory.component.html',
@@ -8,14 +9,57 @@ import * as $ from 'jquery'
 })
 export class InventoryComponent implements OnInit {
 
-  constructor(private PS:ProductsService) { }
+  productForm:FormGroup
+
+  constructor(private PS:ProductsService , private fb:FormBuilder) {
+    this.productForm = this.fb.group(
+      {
+        id : ['',[Validators.required]],
+        title: ['',[Validators.required]],
+        price: ['',[Validators.required]],
+        category: ['',[Validators.required]],
+        quantity: ['',[Validators.required]],
+        description: ['',[Validators.required]],
+        image: ['']
+      }
+    )
+   }
+
+   get productFormControl() {
+    return this.productForm.controls;
+  }
 
   products:any
 
   searchTerm:any = ''
 
-  ngOnInit(): void {
+  getProductData(){ 
     this.PS.productsData().subscribe(res=>this.products=res)
+  }
+
+  ngOnInit(): void {
+    this.getProductData()
+  }
+
+  
+  submit(){
+    this.productForm.value.image = this.imageArray[0]
+    this.PS.addProduct(this.productForm.value).subscribe(res=>{
+      console.log(res);
+      this.getProductData()
+    })
+  }
+
+  imageArray:string[]=[]
+
+  getImage(e:any){
+   if(e.target.files){
+     var reader = new FileReader();
+     reader.readAsDataURL(e.target.files[0])
+    reader.onload = (event:any)=>{
+      this.imageArray.push(event.target.result)
+    }
+   }
   }
 
 }
