+++ src/app/services/products.service.ts
@@ -12,7 +12,7 @@ export class ProductsService {
   products:any
 
   productsData():Observable<any>{
-    return this.hc.get("http://localhost:5100/products")
+    return this.hc.get("http://oshopping.ddns.net/api/products/")
   }
 
   addProduct(data:any){
