+++ src/app/user/item-display/item-display.component.css
@@ -1,8 +1,8 @@
 .image img{
+    margin-top: 10%;
     width: 70%;
 }
 ngb-rating{
-    
     margin-bottom: 20px;
     font-size: xx-large;
     color: orangered;
